Refactor AppSync components for improved zoom functionality and button states
- Updated `AppSyncScanScreen` to use a float for zoom level instead of an integer, enhancing zoom precision.
- Modified `AppsyncButton` to accept an `enabled` parameter, allowing for dynamic button states based on zoom capabilities.
- Enhanced `OverlayControls` to calculate and manage enabled states for zoom in and zoom out buttons, improving user experience.
- Added previews for disabled button states and various zoom levels in `OverlayControls` for better UI testing.

+++ Android/app/appsync/src/main/java/com/greatergoods/libs/appsync/screen/AppSyncScanScreen.kt
@@ -188,7 +188,7 @@ fun AppSyncScanScreen(
                     },
                 )
                 OverlayControls(
-                    zoomLevel = zoomLevel.toInt(),
+                    zoomLevel = zoomLevel,
                     onZoomIn = {
                         if (zoomLevel < AppSyncConstants.MAX_ZOOM) {
                             zoomLevel += AppSyncConstants.ZOOM_STEP
